--- a/jest.config.ts
+++ b/jest.config.ts
@@ -4,12 +4,12 @@ import type { Config } from "@jest/types";
 // Or async function
 export default async (): Promise<Config.InitialOptions> => {
   return {
-    preset: "ts-jest",
     testEnvironment: "jsdom",
     testMatch: ["<rootDir>/src/__tests__/**/*.(spec|test).(ts|tsx)"],
     verbose: true,
     transform: {
-      "\\.[jt]sx?$": "esbuild-jest",
+      "^.+\\.[jt]sx?$": "esbuild-jest",
     },
+    coverageProvider: "v8",
   };
 };