--- a/package.json
+++ b/package.json
@@ -154,7 +154,6 @@
     "rollup-plugin-typescript2": "^0.31.0",
     "shelljs": "0.8.5",
     "terser": "^5.6.1",
-    "ts-jest": "^26.5.5",
     "ts-node": "^9.1.1",
     "tslib": "^2.3.0",
     "typescript": "^4.3.5",