Full test time:

`babel-jest`:
```
Test Suites: 1 skipped, 56 passed, 56 of 57 total
Tests:       10 skipped, 261 passed, 271 total
Snapshots:   0 total
Time:        17.344 s, estimated 20 s
```
`esbuild-jest`:
```
Test Suites: 1 skipped, 56 passed, 56 of 57 total
Tests:       10 skipped, 261 passed, 271 total
Snapshots:   0 total
Time:        10.628 s, estimated 12 s
```